get Rooms and Facilities for making a reservation
