+++ customer_app/actions/deleteActions.jsx
@@ -1,0 +1,7 @@
+// export const deleteFacility = (axios, id) => {
+//   return axios.delete(`facility/admin/delete?id=${id}`)
+// }
+
+// export const deleteRoom = (axios, id) => {
+//   return axios.delete(`room/admin/delete?id=${id}`)
+// }
