+++ customer_app/actions/postActions.jsx
@@ -1,0 +1,7 @@
+// export const addNewFacility = async (axios, facilityData) => {
+//   return axios.post("facility/admin/add", facilityData)
+// }
+
+// export const addNewRoom = async (axios, roomData) => {
+//   return axios.post("room/admin/add", roomData)
+// }
