+++ customer_app/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@nextui-org/react": "^2.3.6",
     "@tanstack/react-query": "^5.37.1",
+    "axios": "^1.7.2",
     "framer-motion": "^11.2.6",
     "qrcode.react": "^3.1.0",
     "react": "^18.3.1",
