+++ customer_app/src/main.jsx
@@ -3,13 +3,18 @@ import ReactDOM from 'react-dom/client'
 import { NextUIProvider } from '@nextui-org/react'
 import App from './App.jsx'
 import './index.css'
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+
+const queryClient = new QueryClient();
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <NextUIProvider>
-      <main className='light text-foreground bg-background'>
-        <App />
-      </main>
-    </NextUIProvider>
+    <QueryClientProvider client={queryClient}>
+      <NextUIProvider>
+        <main className='light text-foreground bg-background'>
+          <App />
+        </main>
+      </NextUIProvider>
+    </QueryClientProvider>
   </React.StrictMode>
 )
